move is_clifford_angle and is_close_angle
